Fix linting warnings and optimize code structure
- Remove unused variables in API files (quantity, location)
- Remove unnecessary eslint-disable directive
- Refactor seoUtils to separate utility functions from React components
- Move utility functions to seoUtils.ts to resolve react-refresh warnings
- All linting warnings resolved, build passes successfully
